pyproject.toml
[project]
name = "aqualungs"
version = "0.1.0"
description = "Collect and digest arXiv.org computer science articles"
authors = [
    {name = "datasuspense",email = "datasuspense@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "omegaconf (>=2.3.0,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "tqdm (>=4.67.3,<5.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "google-api-python-client (>=2.190.0,<3.0.0)",
    "google-auth-httplib2 (>=0.3.0,<0.4.0)",
    "google-auth-oauthlib (>=1.3.0,<2.0.0)",
    "openai (>=2.24.0,<3.0.0)",
    "pdfplumber (>=0.11.9,<0.12.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
